CTCP-6207: Added AsyncCacheApi to reference data connector. MOved connector test to it package.
