+++ build.sbt
@@ -1,23 +1,23 @@
 import play.sbt.routes.RoutesKeys
-import sbt.Def
 import scoverage.ScoverageKeys
 import uk.gov.hmrc.DefaultBuildSettings
 import uk.gov.hmrc.versioning.SbtGitVersioning.autoImport.majorVersion
 
 lazy val appName: String = "manage-transit-movements-frontend"
 
-lazy val root = (project in file("."))
+ThisBuild / majorVersion := 0
+ThisBuild / scalaVersion := "3.5.0"
+ThisBuild / scalafmtOnCompile := true
+
+lazy val microservice = (project in file("."))
   .enablePlugins(PlayScala, SbtDistributablesPlugin, ScalaxbPlugin)
   .disablePlugins(JUnitXmlReportPlugin)
   .settings(DefaultBuildSettings.scalaSettings *)
   .settings(DefaultBuildSettings.defaultSettings() *)
-  .settings(inConfig(Test)(testSettings) *)
   .configs(A11yTest)
   .settings(inConfig(A11yTest)(org.scalafmt.sbt.ScalafmtPlugin.scalafmtConfigSettings) *)
   .settings(headerSettings(A11yTest) *)
   .settings(automateHeaderSettings(A11yTest))
-  .settings(majorVersion := 0)
-  .settings(scalaVersion := "3.5.0")
   .settings(
     Compile / scalaxb / scalaxbXsdSource := new File("./conf/xsd"),
     Compile / scalaxb / scalaxbDispatchVersion := "1.1.3",
@@ -62,14 +62,13 @@ lazy val root = (project in file("."))
       Resolver.jcenterRepo
     ),
     Assets / pipelineStages        := Seq(digest),
-    ThisBuild / useSuperShell      := false,
-    ThisBuild / scalafmtOnCompile := true
+    ThisBuild / useSuperShell      := false
   )
 
-lazy val testSettings: Seq[Def.Setting[?]] = Seq(
-  fork := true,
-  javaOptions ++= Seq(
-    "-Dconfig.resource=test.application.conf",
-    "-Dlogger.resource=logback-test.xml"
+lazy val it = project
+  .enablePlugins(PlayScala)
+  .dependsOn(microservice % "test->test") // the "test->test" allows reusing test code and test dependencies
+  .settings(
+    libraryDependencies ++= AppDependencies.test,
+    DefaultBuildSettings.itSettings()
   )
-)
