+++ conf/application.conf
@@ -104,6 +104,10 @@ tracking-consent-frontend {
   gtm.container = "b"
 }
 
+async-cache-api {
+  expiration = 5
+}
+
 session {
   timeoutSeconds = 3600
   countdownSeconds = 120
