+++ it/test/itbase/ItSpecBase.scala
@@ -1,0 +1,59 @@
+/*
+ * Copyright 2024 HM Revenue & Customs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package itbase
+
+import config.{PostTransitionModule, TransitionModule}
+import models.{DepartureId, LocalReferenceNumber}
+import org.scalatest.OptionValues
+import org.scalatest.concurrent.{IntegrationPatience, ScalaFutures}
+import org.scalatest.freespec.AnyFreeSpec
+import org.scalatest.matchers.must.Matchers
+import org.scalatestplus.play.guice.GuiceOneServerPerSuite
+import play.api.Application
+import play.api.inject.guice.GuiceApplicationBuilder
+import uk.gov.hmrc.http.HeaderCarrier
+
+trait ItSpecBase extends AnyFreeSpec with Matchers with ScalaFutures with OptionValues with GuiceOneServerPerSuite with IntegrationPatience {
+
+  val lrn: LocalReferenceNumber = LocalReferenceNumber("ABCD1234567890123")
+  val mrn: String               = "ABCD1234567890123"
+  val departureId: DepartureId  = DepartureId(1)
+  val messageId: String         = "343ffafafaaf"
+  val departureIdP5: String     = "643cffea2dca70b2"
+  val arrivalIdP5: String       = "62f4ebbb765ba8c2"
+
+  implicit val hc: HeaderCarrier = HeaderCarrier()
+
+  def guiceApplicationBuilder(): GuiceApplicationBuilder =
+    new GuiceApplicationBuilder()
+      .configure("metrics.enabled" -> false)
+
+  final override def fakeApplication(): Application =
+    guiceApplicationBuilder()
+      .build()
+
+  protected def transitionApplicationBuilder(): GuiceApplicationBuilder =
+    guiceApplicationBuilder()
+      .disable[PostTransitionModule]
+      .bindings(new TransitionModule)
+
+  protected def postTransitionApplicationBuilder(): GuiceApplicationBuilder =
+    guiceApplicationBuilder()
+      .disable[TransitionModule]
+      .bindings(new PostTransitionModule)
+
+}
