+++ it/test/itbase/WireMockServerHandler.scala
@@ -14,21 +14,20 @@
  * limitations under the License.
  */
 
-package helper
+package itbase
 
 import com.github.tomakehurst.wiremock.WireMockServer
-import org.scalatest.BeforeAndAfterAll
-import org.scalatest.BeforeAndAfterEach
-import org.scalatest.Suite
 import com.github.tomakehurst.wiremock.core.WireMockConfiguration.wireMockConfig
+import org.scalatest.{BeforeAndAfterAll, BeforeAndAfterEach, Suite}
 
 trait WireMockServerHandler extends BeforeAndAfterAll with BeforeAndAfterEach {
   this: Suite =>
 
-  protected val server: WireMockServer = {
-    val svr = new WireMockServer(wireMockConfig.dynamicPort())
-    svr.start()
-    svr
+  protected val server: WireMockServer = new WireMockServer(wireMockConfig.dynamicPort())
+
+  override def beforeAll(): Unit = {
+    server.start()
+    super.beforeAll()
   }
 
   override def beforeEach(): Unit = {
@@ -40,5 +39,4 @@ trait WireMockServerHandler extends BeforeAndAfterAll with BeforeAndAfterEach {
     super.afterAll()
     server.stop()
   }
-
 }
